backend/requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
Pillow
openai
python-dotenv
